let editor extensions surface monaco field editors (#10511)
